Switched filebase to crc32.
It's ok to use crc32 and it lowers the dependencies. Having another hash algorithm just for that seems to be an overkill.

+++ crates/dizbase/src/file_base/file_info.rs
@@ -32,7 +32,7 @@ impl FileInfo {
         self
     }
 
-    pub fn with_hash(mut self, hash: u64) -> Self {
+    pub fn with_hash(mut self, hash: u32) -> Self {
         self.header.hash = hash;
         self
     }
